Conflicts:
      .nongit

+++ src/storage/browser/quota/quota_database.cc
@@ -61,9 +61,11 @@ const char kBucketTable[] = "buckets";
 // TODO(crbug.com/1254535): Remove once enough time has passed to ensure that
 // this flag is no longer stored and supported in the QuotaDatabase.
 const char kIsOriginTableBootstrapped[] = "IsOriginTableBootstrapped";
+// Deprecated bootstrap flag, invalidated in 03/2022 as part of crbug/1306279.
+const char kDeprecatedBucketsTableBootstrapped[] = "IsBucketsTableBootstrapped";
 // Flag to ensure that all existing data for storage keys have been
 // registered into the buckets table.
-const char kBucketsTableBootstrapped[] = "IsBucketsTableBootstrapped";
+const char kBucketsTableBootstrapped[] = "IsBucketsBootstrapped";
 
 const int kCommitIntervalMs = 30000;
 
@@ -199,12 +201,11 @@ QuotaError QuotaDatabase::SetHostQuota(const std::string& host,
 }
 
 QuotaErrorOr<BucketInfo> QuotaDatabase::GetOrCreateBucket(
-    const StorageKey& storage_key,
-    const std::string& bucket_name) {
+    const BucketInitParams& params) {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
 
   QuotaErrorOr<BucketInfo> bucket_result =
-      GetBucket(storage_key, bucket_name, StorageType::kTemporary);
+      GetBucket(params.storage_key, params.name, StorageType::kTemporary);
 
   if (bucket_result.ok())
     return bucket_result;
@@ -213,8 +214,9 @@ QuotaErrorOr<BucketInfo> QuotaDatabase::GetOrCreateBucket(
     return bucket_result.error();
 
   base::Time now = base::Time::Now();
-  return CreateBucketInternal(storage_key, StorageType::kTemporary, bucket_name,
-                              /*use_count=*/0, now, now);
+  return CreateBucketInternal(
+      params.storage_key, StorageType::kTemporary, params.name,
+      /*use_count=*/0, now, now, params.expiration, params.quota);
 }
 
 QuotaErrorOr<BucketInfo> QuotaDatabase::GetOrCreateBucketDeprecated(
@@ -234,7 +236,7 @@ QuotaErrorOr<BucketInfo> QuotaDatabase::GetOrCreateBucketDeprecated(
 
   base::Time now = base::Time::Now();
   return CreateBucketInternal(storage_key, type, bucket_name, /*use_count=*/0,
-                              now, now);
+                              now, now, absl::nullopt, 0);
 }
 
 QuotaErrorOr<BucketInfo> QuotaDatabase::CreateBucketForTesting(
@@ -245,7 +247,7 @@ QuotaErrorOr<BucketInfo> QuotaDatabase::CreateBucketForTesting(
 
   base::Time now = base::Time::Now();
   return CreateBucketInternal(storage_key, storage_type, bucket_name,
-                              /*use_count=*/0, now, now);
+                              /*use_count=*/0, now, now, absl::nullopt, 0);
 }
 
 QuotaErrorOr<BucketInfo> QuotaDatabase::GetBucket(
@@ -558,16 +560,21 @@ QuotaError QuotaDatabase::DeleteHostQuota(const std::string& host,
   return QuotaError::kNone;
 }
 
-QuotaError QuotaDatabase::DeleteBucketInfo(BucketId bucket_id) {
+QuotaError QuotaDatabase::DeleteBucketData(const BucketLocator& bucket) {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
-  DCHECK(!bucket_id.is_null());
   QuotaError open_error = EnsureOpened();
   if (open_error != QuotaError::kNone)
     return open_error;
 
+  // Doom bucket directory first so data is no longer accessible, even if
+  // directory deletion fails. `storage_directory_` may be nullptr for
+  // in-memory only.
+  if (storage_directory_ && !storage_directory_->DoomBucket(bucket))
+    return QuotaError::kFileOperationError;
+
   static constexpr char kSql[] = "DELETE FROM buckets WHERE id = ?";
   sql::Statement statement(db_->GetCachedStatement(SQL_FROM_HERE, kSql));
-  statement.BindInt64(0, bucket_id.value());
+  statement.BindInt64(0, bucket.id.value());
 
   if (!statement.Run())
     return QuotaError::kDatabaseError;
@@ -584,6 +591,9 @@ QuotaError QuotaDatabase::DeleteBucketInfo(BucketId bucket_id) {
   // TODO(crbug.com/1314567): For handling inconsistencies between the db and
   // the file system.
   ScheduleCommit();
+
+  if (storage_directory_)
+    storage_directory_->ClearDoomedBuckets();
 
   return QuotaError::kNone;
 }
@@ -709,6 +719,7 @@ QuotaError QuotaDatabase::SetIsBootstrapped(bool bootstrap_flag) {
   // TODO(crbug.com/1254535): Remove once enough time has passed to ensure that
   // this flag is no longer stored and supported in the QuotaDatabase.
   meta_table_->DeleteKey(kIsOriginTableBootstrapped);
+  meta_table_->DeleteKey(kDeprecatedBucketsTableBootstrapped);
 
   return meta_table_->SetValue(kBucketsTableBootstrapped, bootstrap_flag)
              ? QuotaError::kNone
@@ -719,8 +730,11 @@ QuotaError QuotaDatabase::RazeAndReopen() {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
   // Try creating a database one last time if there isn't one.
   if (!db_) {
-    if (!db_file_path_.empty())
+    if (!db_file_path_.empty()) {
+      DCHECK(!legacy_db_file_path_.empty());
       sql::Database::Delete(db_file_path_);
+      sql::Database::Delete(legacy_db_file_path_);
+    }
     return EnsureOpened();
   }
 
@@ -818,9 +832,15 @@ QuotaError QuotaDatabase::EnsureOpened() {
       }));
 
   // Migrate an existing database from the old path.
-  if (!db_file_path_.empty()) {
-    if (!MoveLegacyDatabase())
+  if (!db_file_path_.empty() && !MoveLegacyDatabase()) {
+    if (!ResetStorage()) {
+      is_disabled_ = true;
+      db_.reset();
+      meta_table_.reset();
       return QuotaError::kDatabaseError;
+    }
+    // ResetStorage() has succeeded and database is already open.
+    return QuotaError::kNone;
   }
 
   if (!OpenDatabase() || !EnsureDatabaseVersion()) {
@@ -843,22 +863,27 @@ QuotaError QuotaDatabase::EnsureOpened() {
 bool QuotaDatabase::MoveLegacyDatabase() {
   // Migration was added on 04/2022 (https://crrev.com/c/3513545).
   // Cleanup after enough time has passed.
-  if (!base::PathExists(legacy_db_file_path_))
+  if (base::PathExists(db_file_path_) ||
+      !base::PathExists(legacy_db_file_path_)) {
     return true;
+  }
 
-  DCHECK(!base::PathExists(db_file_path_));
   if (!base::CreateDirectory(db_file_path_.DirName()) ||
-      !base::Move(legacy_db_file_path_, db_file_path_)) {
+      !base::CopyFile(legacy_db_file_path_, db_file_path_)) {
+    sql::Database::Delete(db_file_path_);
     return false;
   }
 
   base::FilePath legacy_journal_path =
       sql::Database::JournalPath(legacy_db_file_path_);
   if (base::PathExists(legacy_journal_path) &&
-      !base::Move(legacy_journal_path,
-                  sql::Database::JournalPath(db_file_path_))) {
+      !base::CopyFile(legacy_journal_path,
+                      sql::Database::JournalPath(db_file_path_))) {
+    sql::Database::Delete(db_file_path_);
     return false;
   }
+
+  sql::Database::Delete(legacy_db_file_path_);
   return true;
 }
 
@@ -996,6 +1021,12 @@ bool QuotaDatabase::ResetStorage() {
   sql::Database::Delete(legacy_db_file_path_);
   sql::Database::Delete(db_file_path_);
 
+  // Explicit file deletion to try and get consistent deletion across platforms.
+  base::DeleteFile(legacy_db_file_path_);
+  base::DeleteFile(db_file_path_);
+  base::DeleteFile(sql::Database::JournalPath(legacy_db_file_path_));
+  base::DeleteFile(sql::Database::JournalPath(db_file_path_));
+
   storage_directory_->Doom();
   storage_directory_->ClearDoomed();
 
@@ -1050,7 +1081,9 @@ QuotaErrorOr<BucketInfo> QuotaDatabase::CreateBucketInternal(
     const std::string& bucket_name,
     int use_count,
     base::Time last_accessed,
-    base::Time last_modified) {
+    base::Time last_modified,
+    absl::optional<base::Time> expiration,
+    int64_t quota) {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
   // TODO(crbug/1210259): Add DCHECKs for input validation.
   QuotaError open_error = EnsureOpened();
@@ -1069,7 +1102,7 @@ QuotaErrorOr<BucketInfo> QuotaDatabase::CreateBucketInternal(
         "last_modified,"
         "expiration,"
         "quota) "
-        "VALUES (?, ?, ?, ?, ?, ?, ?, ?, 0)";
+        "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)";
   // clang-format on
   sql::Statement statement(db_->GetCachedStatement(SQL_FROM_HERE, kSql));
   statement.BindString(0, storage_key.Serialize());
@@ -1079,7 +1112,9 @@ QuotaErrorOr<BucketInfo> QuotaDatabase::CreateBucketInternal(
   statement.BindInt(4, use_count);
   statement.BindTime(5, last_accessed);
   statement.BindTime(6, last_modified);
-  statement.BindTime(7, base::Time::Max());
+  const base::Time expires = expiration.value_or(base::Time::Max());
+  statement.BindTime(7, expires);
+  statement.BindInt64(8, quota);
 
   if (!statement.Run())
     return QuotaError::kDatabaseError;
@@ -1093,7 +1128,7 @@ QuotaErrorOr<BucketInfo> QuotaDatabase::CreateBucketInternal(
   Commit();
 
   return BucketInfo(BucketId(bucket_id), storage_key, type, bucket_name,
-                    base::Time::Max(), 0);
+                    expires, quota);
 }
 
 bool operator<(const QuotaDatabase::BucketTableEntry& lhs,
